Shichao Versions
Here are Shichao's versions of the files as he sent them to me this afternoon (3/12).  Note that css.m runs an SPGL1 algorithm, and the variable 'nn' is the sample subset dimension.  Also, the cq_realtime files are copies of those from Josh Goings's git repo of the same name corresponding to simulated H2O dipole data (the .csv files)